add page_directory from u32

+++ srcs/paging/mod.rs
@@ -11,5 +11,5 @@ macro_rules! enable_paging {
 		"mov eax, cr0",
 		"or eax, 0x80000001",
 		"mov cr0, eax",
-		p = in(reg) &$page_directory as *const _)};);
+		p = in(reg) (&$page_directory as *const _) as usize)};);
 }
